+++ packages/trustix-nix-reprod/internal/cmd/index-logs.go
