+++ packages/trustix-nix-reprod/internal/eval/interface.go
@@ -10,7 +10,7 @@ package eval
 
 import (
 	"context"
-	"github.com/nix-community/trustix/packages/trustix-nix-reprod/lib"
+	"github.com/nix-community/trustix/packages/trustix-nix-reprod/internal/lib"
 )
 
 type Evaluator func(context.Context, *EvalConfig) (chan *lib.Result[*EvalResult], error)
