+++ packages/trustix-nix-reprod/sqlc.yaml
@@ -3,7 +3,7 @@ project:
     id: ""
 packages:
   - name: "db"
-    path: "db"
+    path: "internal/db"
     engine: "sqlite"
     schema: "sql/schema"
     queries: "sql/queries"
